First prove-of-concept implementation of async mqttc client
The idea is to build a new layer on top of the eventloop that provides
the user a very simple and convenient interface. Using the eventloop
directly gives more control but at the cost of having to match
responses to requests manually. The idea behind this client is
instead that e.g. if client.subscribe(...).await is called, this
call blocks (asynchronously) until the corresponding suback is
received. For publishing calls the blocking duration depends on the
chosen QoS. For AtMostOnce it should block until the message is
written to the TCP/IP stream. For higher QoS it should block until
all the necessary mqtt control messages are exchanged.

At this time, the matching of requests to responses is NOT yet
correctly implemented. It's more a first design attempt for the
client interface.